# Conflicts:
#	osmdroid-android/src/main/java/org/osmdroid/tileprovider/modules/SqlTileWriter.java

+++ osmdroid-android/src/main/java/org/osmdroid/tileprovider/modules/SqlTileWriter.java
@@ -3,6 +3,7 @@ package org.osmdroid.tileprovider.modules;
 import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteFullException;
 import android.util.Log;
 
 import org.osmdroid.api.IMapView;
@@ -21,6 +22,7 @@ import java.util.List;
 
 import static org.osmdroid.tileprovider.modules.DatabaseFileArchive.COLUMN_PROVIDER;
 import static org.osmdroid.tileprovider.modules.DatabaseFileArchive.COLUMN_TILE;
+import static org.osmdroid.tileprovider.modules.DatabaseFileArchive.COLUMN_KEY;
 import static org.osmdroid.tileprovider.modules.DatabaseFileArchive.TABLE;
 
 /**
@@ -35,14 +37,16 @@ import static org.osmdroid.tileprovider.modules.DatabaseFileArchive.TABLE;
  * @since 5.1
  */
 public class SqlTileWriter implements IFilesystemCache {
+    public static final String DATABASE_FILENAME = "cache.db";
+
     protected File db_file;
     protected SQLiteDatabase db;
-    final int questimate=8000;
+    final int questimate=4000;
     static boolean hasInited=false;
 
     public SqlTileWriter() {
 
-        db_file = new File(OpenStreetMapTileProviderConstants.TILE_PATH_BASE.getAbsolutePath() + File.separator + "cache.db");
+        db_file = new File(OpenStreetMapTileProviderConstants.TILE_PATH_BASE.getAbsolutePath() + File.separator + DATABASE_FILENAME);
         OpenStreetMapTileProviderConstants.TILE_PATH_BASE.mkdirs();
 
         try {
@@ -59,48 +63,59 @@ public class SqlTileWriter implements IFilesystemCache {
             final Thread t = new Thread() {
                 @Override
                 public void run() {
-                    if (db==null) {
-                        if (OpenStreetMapTileProviderConstants.DEBUGMODE) {
-                            Log.d(IMapView.LOGTAG, "Finished init thread, aborted due to null database reference");
-                        }
-                        return;
-                    }
-
-                    try {
-                        //run the reaper (remove all old expired tiles)
-                        //keep if now is < expiration date
-                        //delete if now is > expiration date
-                        long now = System.currentTimeMillis();
-                        int rows = db.delete(TABLE, "expires < ?", new String[]{System.currentTimeMillis() + ""});
-                        Log.d(IMapView.LOGTAG, "Local storage cahce purged " + rows + " expired tiles in " + (System.currentTimeMillis() - now) + "ms, cache size is " + db_file.length() + "bytes");
-
-                        //VACUUM the database
-                        now = System.currentTimeMillis();
-                        //db.execSQL("VACUUM " + DatabaseFileArchive.TABLE + ";");
-                        // Log.d(IMapView.LOGTAG, "VACUUM completed in " + (System.currentTimeMillis()-now) + "ms, cache size is " + db_file.length() + "bytes");
-                        if (db_file.length() > OpenStreetMapTileProviderConstants.TILE_MAX_CACHE_SIZE_BYTES) {
-                            long diff = OpenStreetMapTileProviderConstants.TILE_MAX_CACHE_SIZE_BYTES - db_file.length();
-                            long tilesToKill = diff / questimate;
-                            try {
-                                db.execSQL("DELETE FROM " + TABLE + " ORDER BY expires DESC LIMIT " + tilesToKill);
-                            } catch (Throwable t) {
-                                t.printStackTrace();
-                            }
-                            Log.d(IMapView.LOGTAG, "purge completed in " + (System.currentTimeMillis() - now) + "ms, cache size is " + db_file.length() + "bytes");
-                        }
-                    }catch (Exception ex){
-                        if (OpenStreetMapTileProviderConstants.DEBUGMODE) {
-                            Log.d(IMapView.LOGTAG, "SqliteTileWriter init thread crash, db is probably not available",ex);
-                        }
-                    }
-
-                    if (OpenStreetMapTileProviderConstants.DEBUGMODE) {
-                        Log.d(IMapView.LOGTAG, "Finished init thread");
-                    }
+                    runCleanupOperation();
                 }
             };
             t.setPriority(Thread.MIN_PRIORITY);
             t.start();
+        }
+    }
+
+    /**
+     * this could be a long running operation, don't run on the UI thread unless necessary.
+     * This function prunes the database for old or expired tiles.
+     */
+    public void runCleanupOperation() {
+        if (db==null) {
+            if (OpenStreetMapTileProviderConstants.DEBUGMODE) {
+                Log.d(IMapView.LOGTAG, "Finished init thread, aborted due to null database reference");
+            }
+            return;
+        }
+
+        try {
+            //run the reaper (remove all old expired tiles)
+            //keep if now is < expiration date
+            //delete if now is > expiration date
+            long now = System.currentTimeMillis();
+            int rows = db.delete(TABLE, "expires < ?", new String[]{System.currentTimeMillis() + ""});
+            Log.d(IMapView.LOGTAG, "Local storage cache purged " + rows + " expired tiles in " + (System.currentTimeMillis() - now) + "ms, cache size is " + db_file.length() + "bytes");
+
+            //VACUUM the database
+            Log.d(IMapView.LOGTAG, "Local cache is now " + db_file.length() + " max size is " + OpenStreetMapTileProviderConstants.TILE_MAX_CACHE_SIZE_BYTES);
+            now = System.currentTimeMillis();
+            //db.execSQL("VACUUM " + DatabaseFileArchive.TABLE + ";");
+            // Log.d(IMapView.LOGTAG, "VACUUM completed in " + (System.currentTimeMillis()-now) + "ms, cache size is " + db_file.length() + "bytes");
+            if (db_file.length() > OpenStreetMapTileProviderConstants.TILE_MAX_CACHE_SIZE_BYTES) {
+                long diff = db_file.length() - OpenStreetMapTileProviderConstants.TILE_MAX_CACHE_SIZE_BYTES ;
+                long tilesToKill = diff / questimate;
+                Log.d(IMapView.LOGTAG, "Local cache purging " +tilesToKill + " tiles.");
+                if (tilesToKill > 0)
+                try {
+                    db.execSQL("DELETE FROM " + TABLE + " WHERE " + COLUMN_KEY + " in (SELECT " + COLUMN_KEY + " FROM " + TABLE + " ORDER BY expires DESC LIMIT " + tilesToKill + ")");
+                } catch (Throwable t) {
+                    Log.e(IMapView.LOGTAG, "error purging tiles from the tile cache", t);
+                }
+                Log.d(IMapView.LOGTAG, "purge completed in " + (System.currentTimeMillis() - now) + "ms, cache size is " + db_file.length() + " bytes");
+            }
+        }catch (Exception ex){
+            if (OpenStreetMapTileProviderConstants.DEBUGMODE) {
+                Log.d(IMapView.LOGTAG, "SqliteTileWriter init thread crash, db is probably not available",ex);
+            }
+        }
+
+        if (OpenStreetMapTileProviderConstants.DEBUGMODE) {
+            Log.d(IMapView.LOGTAG, "Finished init thread");
         }
     }
 
@@ -134,9 +149,15 @@ public class SqlTileWriter implements IFilesystemCache {
             //this shouldn't happen, but just in case
             if (pTile.getExpires() != null)
                 cv.put("expires", pTile.getExpires().getTime());
-            db.delete(TABLE, DatabaseFileArchive.COLUMN_KEY+"=? and "+DatabaseFileArchive.COLUMN_PROVIDER+"=?", new String[]{index+"",pTileSourceInfo.name()});
+            db.delete(TABLE, DatabaseFileArchive.COLUMN_KEY + "=? and " + DatabaseFileArchive.COLUMN_PROVIDER + "=?", new String[]{index + "", pTileSourceInfo.name()});
             db.insert(TABLE, null, cv);
-            Log.d(IMapView.LOGTAG, "tile inserted " + pTileSourceInfo.name() +  pTile.toString());
+            Log.d(IMapView.LOGTAG, "tile inserted " + pTileSourceInfo.name() + pTile.toString());
+            if (db_file.length() > OpenStreetMapTileProviderConstants.TILE_TRIM_CACHE_SIZE_BYTES){
+                runCleanupOperation();
+            }
+        } catch (SQLiteFullException ex) {
+            //the drive is full! trigger the clean up operation
+            runCleanupOperation();
         } catch (Throwable ex) {
             //note, although we check for db null state at the beginning of this method, it's possible for the
             //db to be closed during the execution of this method
@@ -146,15 +167,14 @@ public class SqlTileWriter implements IFilesystemCache {
         return false;
     }
 
-    @Override
-    public boolean exists(ITileSource pTileSource, MapTile pTile) {
+    public boolean exists(String pTileSource, MapTile pTile) {
         try {
             final String[] tile = {DatabaseFileArchive.COLUMN_TILE};
             final long x = (long) pTile.getX();
             final long y = (long) pTile.getY();
             final long z = (long) pTile.getZoomLevel();
             final long index = ((z << z) + x << z) + y;
-            final Cursor cur = db.query(TABLE, tile, DatabaseFileArchive.COLUMN_KEY+" = " + index + " and "+DatabaseFileArchive.COLUMN_PROVIDER+" = '" + pTileSource.name() + "'", null, null, null, null);
+            final Cursor cur = db.query(TABLE, tile, DatabaseFileArchive.COLUMN_KEY+" = " + index + " and "+DatabaseFileArchive.COLUMN_PROVIDER+" = '" + pTileSource + "'", null, null, null, null);
 
             if(cur.getCount() != 0) {
                 cur.close();
@@ -162,9 +182,14 @@ public class SqlTileWriter implements IFilesystemCache {
             }
             cur.close();
         } catch (Throwable ex) {
-            Log.e(IMapView.LOGTAG, "Unable to store cached tile from " + pTileSource.name() + " " + pTile.toString(), ex);
+            Log.e(IMapView.LOGTAG, "Unable to store cached tile from " + pTileSource + " " + pTile.toString(), ex);
         }
         return false;
+    }
+
+    @Override
+    public boolean exists(ITileSource pTileSource, MapTile pTile) {
+        return exists(pTileSource.name(), pTile);
     }
 
     @Override
@@ -197,7 +222,9 @@ public class SqlTileWriter implements IFilesystemCache {
 
     /**
      * a helper method to import file system stored map tiles into the sql tile cache
-     * on successful import, the tiles are removed from the file system
+     * on successful import, the tiles are removed from the file system.
+     *
+     * This can take a long time, so consider running this off of the main thread.
      * @return
      */
     public int[] importFromFileCache(boolean removeFromFileSystem) {
@@ -233,34 +260,40 @@ public class SqlTileWriter implements IFilesystemCache {
                                                                 final long z1 = Long.parseLong(z[zz].getName());
                                                                 final long index = ((z1 << z1) + x1 << z1) + y1;
                                                                 cv.put(DatabaseFileArchive.COLUMN_PROVIDER, tileSources[i].getName());
-                                                                BufferedInputStream bis = new BufferedInputStream(new FileInputStream(y[yy]));
+                                                                if (!exists(tileSources[i].getName(), new MapTile((int)z1, (int)x1, (int)y1))) {
 
-                                                                List<Byte> list = new ArrayList<Byte>();
-                                                                //ByteArrayBuffer baf = new ByteArrayBuffer(500);
-                                                                int current = 0;
-                                                                while ((current = bis.read()) != -1) {
-                                                                    list.add((byte) current);
-                                                                }
+                                                                    BufferedInputStream bis = new BufferedInputStream(new FileInputStream(y[yy]));
 
-                                                                byte[] bits = new byte[list.size()];
-                                                                for (int bi = 0; bi < list.size(); bi++) bits[bi] = list.get(bi);
-                                                                cv.put(DatabaseFileArchive.COLUMN_KEY, index);
-                                                                cv.put(DatabaseFileArchive.COLUMN_TILE, bits);
-
-                                                                long insert = db.insert(TABLE, null, cv);
-                                                                if (insert>0) {
-                                                                    Log.d(IMapView.LOGTAG, "tile inserted " + tileSources[i].getName() + "/" + z1 + "/" + x1 + "/" + y1);
-                                                                    ret[0]++;
-                                                                    if (removeFromFileSystem){
-                                                                        try {
-                                                                            y[yy].delete();
-                                                                            ret[2]++;;
-                                                                        }catch (Exception ex){
-                                                                            ret[3]++;;
-                                                                        }
+                                                                    List<Byte> list = new ArrayList<Byte>();
+                                                                    //ByteArrayBuffer baf = new ByteArrayBuffer(500);
+                                                                    int current = 0;
+                                                                    while ((current = bis.read()) != -1) {
+                                                                        list.add((byte) current);
                                                                     }
-                                                                } else {
-                                                                    Log.w(IMapView.LOGTAG, "tile NOT inserted " + tileSources[i].getName() + "/" + z1 + "/" + x1 + "/" + y1);
+
+                                                                    byte[] bits = new byte[list.size()];
+                                                                    for (int bi = 0; bi < list.size(); bi++)
+                                                                        bits[bi] = list.get(bi);
+                                                                    cv.put(DatabaseFileArchive.COLUMN_KEY, index);
+                                                                    cv.put(DatabaseFileArchive.COLUMN_TILE, bits);
+
+                                                                    long insert = db.insert(TABLE, null, cv);
+                                                                    if (insert > 0) {
+                                                                        Log.d(IMapView.LOGTAG, "tile inserted " + tileSources[i].getName() + "/" + z1 + "/" + x1 + "/" + y1);
+                                                                        ret[0]++;
+                                                                        if (removeFromFileSystem) {
+                                                                            try {
+                                                                                y[yy].delete();
+                                                                                ret[2]++;
+                                                                                ;
+                                                                            } catch (Exception ex) {
+                                                                                ret[3]++;
+                                                                                ;
+                                                                            }
+                                                                        }
+                                                                    } else {
+                                                                        Log.w(IMapView.LOGTAG, "tile NOT inserted " + tileSources[i].getName() + "/" + z1 + "/" + x1 + "/" + y1);
+                                                                    }
                                                                 }
 
                                                             } catch (Throwable ex) {
